Fork is kinda working with the llvm backend.
When using continue signal, the wait for the continue signal returns -1,
which is different than what the corresponding c code would do. I don't
yet know why this is. Also when we run the stop test first, the continue
test will hang, so this probably has something to do with it...